# api/openapi.yaml
openapi: 3.0.3
info:
  title: flockd local API
  description: |
    The Teraflock node daemon's loopback HTTP surface on `localhost:7777`.

    Two halves:
    - **management** (`/api/v1/*`): the REST+SSE API behind the CLI, TUI,
      embedded web dashboard, and desktop app. Bearer-token auth (the token
      lives at `<data_dir>/local_api_token`, 0600 — same-user access is the
      security model). This spec **generates the daemon's router and types**
      (oapi-codegen, `make gen`): it cannot drift from the implementation.
    - **openai** (`/v1/*`): the OpenAI-compatible inference surface. Wire
      format follows the OpenAI API; endpooints are documented here for
      completeness but their schemas are owned by that external standard.
      Keyless on loopback unless `local_api.require_auth_v1` is set.
  version: "1.0"
servers:
  - url: http://127.0.0.1:7777
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/models:
    get:
      tags: [management]
      operationId: listModels
      summary: Locally installed models
      responses:
        "200":
          description: Cache contents merged with runtime state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModelList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/events:
    get:
      tags: [events]
      operationId: getEvents
      summary: Live daemon events (SSE)
      description: |
        `text/event-stream`. Event types:
        - `status`: a full Status snapshot — every 2s, and immediately on
          governor state transitions and after enrollment.
        - `model_progress`: `{model, received_bytes, total_bytes}` during
          downloads (throttled).
        - `models_changed`: `{model, change}` where change is
          downloaded|loaded|unloaded|default.
        - `model_assignment`: `{model, state, error}` as a coordinator
          placement moves through assigned|downloading|ready|cached|
          declined|failed|evicted.
        - `activity`: one ActivityEvent as it happens (same rows as
          `GET /api/v1/activity`).
        - `update_available`: the Update object when a version check finds
          a newer release (once per discovered version).
        - `log` (only with `?logs=1`): one LogEntry per line as logged.

        EventSource cannot send headers, so this route additionally accepts
        the bearer token as `?token=`. Hand-written handler — documented
        here, excluded from codegen.
      parameters:
        - name: logs
          in: query
          required: false
          schema: { type: boolean }
        - name: token
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: SSE stream.
          content:
            text/event-stream:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/chat/completions:
    post:
      tags: [openai]
      operationId: openaiChatCompletions
      summary: OpenAI-compatible chat completions (streaming supported)
      description: |
        Reasoning models: chain-of-thought is returned separately from the
        answer as `reasoning_content` (the DeepSeek / llama.cpp extension
        to the OpenAI shape) — on `choices[].message.reasoning_content` for
        non-streaming responses and on `choices[].delta.reasoning_content`
        for stream chunks — and is absent when the model produced none.
        Reasoning tokens count in `usage.completion_tokens`. Models whose
        thinking llama.cpp cannot parse still return it inline in `content`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200":
          description: "Completion (or SSE stream when `stream: true`). Chat messages/deltas may carry `reasoning_content`."
          content:
            application/json:
              schema: { type: object }
        "503":
          description: The governor refuses inference right now (state in the message).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/completions:
    post:
      tags: [openai]
      operationId: openaiCompletions
      summary: OpenAI-compatible text completions
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200":
          description: Completion.
          content:
            application/json:
              schema: { type: object }
  /v1/embeddings:
    post:
      tags: [openai]
      operationId: openaiEmbeddings
      summary: OpenAI-compatible embeddings
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200":
          description: Embeddings.
          content:
            application/json:
              schema: { type: object }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Contents of `<data_dir>/local_api_token`.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message: { type: string }
            type: { type: string }
...
    ModelRow:
      type: object
      required: [id, size_bytes, pinned, last_used, state, loaded, default, origin]
      properties:
        id: { type: string }
        size_bytes: { type: integer, format: int64 }
        pinned: { type: boolean }
        last_used: { type: string, format: date-time }
        state:
          type: string
          description: >
            assigned | downloading | ready | missing. `assigned` is a
            coordinator placement queued behind other work (nothing on disk
            yet). `missing` means the index knows the model but its file is
            gone from disk (deleted outside the daemon); it no longer counts
            against the budget and loads re-download it.
        loaded: { type: boolean }
        default: { type: boolean }
        path:
          type: string
          description: Absolute artifact path; present when the file exists.
        loaded_mb:
          type: integer
          format: int64
          description: Measured memory footprint; present only while loaded.
        idle_since:
          type: string
          format: date-time
          description: >
            Present while loaded and not serving; the daemon unloads the
            instance after `idle_unload_seconds` (default model exempt).
        origin:
          type: string
          description: >
            Who installed it: `operator` (you, via the app/CLI/config) or
            `mesh` (coordinator placement). The mesh can only evict its
            own; yours are never touched.
        received_bytes:
          type: integer
          format: int64
          description: >
            Live progress summed over every file of the artifact; present
            only while downloading.
        parts_total:
          type: integer
          description: >
            Files in the artifact: the GGUF shards of a sharded model plus
            its mmproj sidecar when it has one. Present only for such
            multi-file models (a plain single-file model omits it). They
            live in `<models dir>/<id>/` under their upstream names;
            `path` is the first shard.
        parts_done:
          type: integer
          description: >
            How many of `parts_total` are downloaded and verified; present
            with it. Shards download one at a time, so the one in flight is
            `parts_done + 1`.
        assignment: { $ref: "#/components/schemas/Assignment" }
    Assignment:
      type: object
      description: >
        Coordinator placement status for this model, when the mesh asked
        for it. Failures and refusals stay visible for ten minutes.
      required: [state, since]
      properties:
        state:
          type: string
          description: >
            assigned | downloading | ready | cached | declined | failed |
            evicted. `cached` is a mesh-placed model that is on disk but
            unloaded (idle unload or memory pressure); the coordinator
            re-sends the assignment to load it when demand returns.
        since: { type: string, format: date-time }
        error:
          type: string
          description: Why it was declined or failed.
    ModelList:
      type: object
      required: [models]
      properties:
        models:
          type: array
          items: { $ref: "#/components/schemas/ModelRow" }
...